*:修改构建镜像镜像名称的判断重复的方式； *：修复镜像中心的查询， *：基础镜像不添加前缀路径

+++ src/main/java/com/bonc/epm/paas/controller/UserController.java
@@ -872,7 +872,7 @@ public class UserController {
         
         Map<String, String> hard = quota.getSpec().getHard();
         hard.put("memory", resource.getRam() + "G"); // 内存
-        hard.put("cpu", Integer.valueOf(resource.getCpu_account())/Integer.valueOf(RATIO_MEMTOCPU) + "");// CPU数量
+        hard.put("cpu", Double.valueOf(resource.getCpu_account())/Double.valueOf(RATIO_MEMTOCPU) + "");// CPU数量
 		// hard.put("pods", resource.getPod_count() + "");//POD数量
 		// hard.put("services", resource.getServer_count() + "");//服务
 		// hard.put("replicationcontrollers", resource.getImage_control() +
@@ -933,7 +933,7 @@ public class UserController {
         try {
             Map<String, String> map = new HashMap<String, String>();
             map.put("memory", resource.getRam() + "G"); // 内存
-            map.put("cpu", Integer.valueOf(resource.getCpu_account())/Integer.valueOf(RATIO_MEMTOCPU) + "");// CPU数量(个)
+            map.put("cpu", Double.valueOf(resource.getCpu_account())/Double.valueOf(RATIO_MEMTOCPU) + "");// CPU数量(个)
             map.put("persistentvolumeclaims", resource.getVol() + "");// 卷组数量
             //map.put("pods", resource.getPod_count() + "");//POD数量
             //map.put("services", resource.getServer_count() + "");//服务
